[FLINK-29823] Use null snapshot id as the default value

+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/file/catalog/AbstractCatalog.java
@@ -59,7 +59,7 @@ public abstract class AbstractCatalog implements Catalog {
             } else if (splits.length == 3) {
                 String table = splits[0];
                 String snapshotTag = splits[1];
-                int snapshotId = Integer.parseInt(splits[2]);
+                Long snapshotId = Long.parseLong(splits[2]);
                 checkState(
                         StringUtils.equalsIgnoreCase(snapshotTag, SNAPSHOT),
                         String.format("Table name can only be %s$snapshot$%s", table, snapshotId));
